hotel_system/requirements.txt
django-suit==0.2.26
